Disabling OCW edition flavour (#15)
'm disabling the OCW edition flavour from being added to the container image for the time being. Upstream Mastodon are making a lot of changes to components and keeping up with it isn't easy.

In order to fix this, I'm going to need to look into a better way of handling the dev and build steps for the flavour. If me removing the flavour temporarily breaks functionality for those who have it set as their default, including myself, I will revert back to the last good container image until I can get everything sorted out.